+++ configs/config.properties
@@ -1,0 +1,2 @@
+driverPath=D:\\Documents\\Appium Drivers\\chromedriver.exe
+filePath=D:\\Documents\\baseFramework\\Screenshots
